pyproject.toml
[tool.poetry]
name = "paprika"
version = "0.1.0"
description = "Fork of Taigabot3, an IRC Bot"
authors = ["Anthony DeDominic <adedomin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22"
pydle = "^0.9.1"
beautifulsoup4 = "^4.8"
python-dateutil = "^2.8"
tomlkit = "^0.5.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
